# openapi.yaml
openapi: 3.1.0
info:
  title: Trustify
  description: Software Supply-Chain Security API
  license:
    name: Apache License, Version 2.0
    identifier: Apache-2.0
  version: 0.4.0-beta.1
paths:
...
  /api/v2/admin/sbom/prune:
    post:
      tags:
      - admin
      summary: Prune SBOMs based on ingestion date
      operationId: pruneSboms
      parameters:
      - name: ingested
        in: query
        description: Number of days ago from current time to prune SBOMs
        required: true
        schema:
          type: integer
          format: int32
          minimum: 0
        style: form
        example: 90
      - name: dry_run
        in: query
        description: If true, only return the list of SBOMs that would be deleted without actually deleting them
        required: true
        schema:
          type: boolean
      - name: batch_size
        in: query
        description: Number of SBOMs to process in a single batch
        required: true
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: max_concurrent
        in: query
        description: Maximum number of concurrent operations
        required: true
        schema:
          type: integer
          minimum: 0
      responses:
        '200':
          description: List of pruned SBOMs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrunedSbomLog'
        '500':
          description: Internal server error
...
  /api/v2/sbom/{id}:
    get:
      tags:
      - sbom
      summary: Get information about an SBOM
      operationId: getSbom
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: Matching SBOM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SbomSummary'
        '404':
          description: The SBOM could not be found
    delete:
      tags:
      - sbom
      summary: Delete an SBOM
      operationId: deleteSbom
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: Matching SBOM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SbomSummary'
        '404':
          description: The SBOM could not be found
...
components:
  schemas:
...
    BasePurlHead:
      type: object
      required:
      - uuid
      - purl
      properties:
        purl:
          $ref: '#/components/schemas/Purl'
          description: The actual base PURL
        uuid:
          type: string
          format: uuid
          description: The ID of the base PURL
...
    Id:
      type: string
      description: |-
        Identifier to a document, prefixed with the ID type.

        Either an internal ID of the document with the `urn:uuid:` scheme. Or using a digest, with the digest prefix. For example, `sha256:`.
      examples:
      - urn:uuid:018123ef-a791-40d8-b62a-f70a350245d4
      - sha256:dc60aeb735c16a71b6fc56e84ddb8193e3a6d1ef0b7e958d77e78fc039a5d04e
...
    Labels:
      type: object
      additionalProperties:
        type: string
    LicenseCategory:
      type: string
      enum:
      - declared
      - concluded
    LicenseInfo:
      type: object
      required:
      - license_name
      - license_type
      properties:
        license_name:
          type: string
        license_type:
          $ref: '#/components/schemas/LicenseCategory'
    LicenseRefMapping:
      type: object
      required:
      - license_id
      - license_name
      properties:
        license_id:
          type: string
        license_name:
          type: string
...
    PrunedSbom:
      type: object
      required:
      - sbom_id
      - authors
      - suppliers
      - data_licenses
      - ingested
      properties:
        authors:
          type: array
          items:
            type: string
        data_licenses:
          type: array
          items:
            type: string
        document_id:
          type:
          - string
          - 'null'
        error:
          type:
          - string
          - 'null'
        ingested:
          type: string
          format: date-time
        published:
          type:
          - string
          - 'null'
          format: date-time
        sbom_id:
          type: string
          format: uuid
        suppliers:
          type: array
          items:
            type: string
    PrunedSbomLog:
      type: object
      required:
      - total
      - successful_total
      - failed_total
      - successful_pruned
      - failed_pruned
      properties:
        failed_pruned:
          type: array
          items:
            $ref: '#/components/schemas/PrunedSbom'
        failed_total:
          type: integer
          format: int64
          minimum: 0
        successful_pruned:
          type: array
          items:
            $ref: '#/components/schemas/PrunedSbom'
        successful_total:
          type: integer
          format: int64
          minimum: 0
        total:
          type: integer
          format: int64
          minimum: 0
    Purl:
      type: string
      format: uri
...
    PurlHead:
      type: object
      required:
      - uuid
      - purl
      properties:
        purl:
          $ref: '#/components/schemas/Purl'
          description: The actual qualified PURL
        uuid:
          type: string
          format: uuid
          description: The ID of the qualified PURL
...
    PurlSummary:
      allOf:
      - $ref: '#/components/schemas/PurlHead'
      - type: object
        required:
        - base
        - version
        - qualifiers
        properties:
          base:
            $ref: '#/components/schemas/BasePurlHead'
          qualifiers:
            type: object
            additionalProperties:
              type: string
            propertyNames:
              type: string
            deprecated: true
          version:
            $ref: '#/components/schemas/VersionedPurlHead'
...
    SbomHead:
      type: object
      required:
      - id
      - labels
      - data_licenses
      - published
      - authors
      - suppliers
      - name
      - number_of_packages
      properties:
        authors:
          type: array
          items:
            type: string
          description: Authors of the SBOM
        data_licenses:
          type: array
          items:
            type: string
        document_id:
          type:
          - string
          - 'null'
        id:
          type: string
        labels:
          $ref: '#/components/schemas/Labels'
        name:
          type: string
        number_of_packages:
          type: integer
          format: int64
          description: The number of packages this SBOM has
          minimum: 0
        published:
          type:
          - string
          - 'null'
          format: date-time
        suppliers:
          type: array
          items:
            type: string
          description: Suppliers of the SBOMs content
...
    SbomPackage:
      type: object
      required:
      - id
      - name
      - purl
      - cpe
      - licenses
      - licenses_ref_mapping
      properties:
        cpe:
          type: array
          items:
            type: string
          description: CPEs identifying the package
        group:
          type:
          - string
          - 'null'
          description: An optional group/namespace for an SBOM package
        id:
          type: string
          description: The SBOM internal ID of a package
        licenses:
          type: array
          items:
            $ref: '#/components/schemas/LicenseInfo'
          description: License info
        licenses_ref_mapping:
          type: array
          items:
            $ref: '#/components/schemas/LicenseRefMapping'
          description: LicenseRef mappings
        name:
          type: string
          description: The name of the package in the SBOM
        purl:
          type: array
          items:
            $ref: '#/components/schemas/PurlSummary'
          description: PURLs identifying the package
        version:
          type:
          - string
          - 'null'
          description: An optional version for an SBOM package
...
    SbomSummary:
      allOf:
      - $ref: '#/components/schemas/SbomHead'
      - $ref: '#/components/schemas/SourceDocument'
      - type: object
        required:
        - described_by
        properties:
          described_by:
            type: array
            items:
              $ref: '#/components/schemas/SbomPackage'
...
    SourceDocument:
      type: object
      required:
      - sha256
      - sha384
      - sha512
      - size
      - ingested
      properties:
        ingested:
          type: string
          format: date-time
          description: The timestamp the document was ingested
        sha256:
          type: string
        sha384:
          type: string
        sha512:
          type: string
        size:
          type: integer
          format: int64
          minimum: 0
...
    VersionedPurlHead:
      type: object
      required:
      - uuid
      - purl
      - version
      properties:
        purl:
          $ref: '#/components/schemas/Purl'
          description: The actual, versioned PURL
        uuid:
          type: string
          format: uuid
          description: The ID of the versioned PURL
        version:
          type: string
          description: The version from the PURL
...